converted test to checker
